split test for store and store factory
